Refactor existing search, part 1
